Revert "Split bm_fullstack into pieces, reuse infrastructure across other microbenchmarks"
