+++ build.yaml
@@ -586,7 +586,7 @@ filegroups:
   headers:
   - test/core/end2end/cq_verifier.h
   - test/core/end2end/fake_resolver.h
-  - test/core/end2end/fixtures/http_proxy_fixture.h
+  - test/core/end2end/fixtures/http_proxy.h
   - test/core/end2end/fixtures/proxy.h
   - test/core/iomgr/endpoint_tests.h
   - test/core/util/debugger_macros.h
@@ -602,7 +602,7 @@ filegroups:
   src:
   - test/core/end2end/cq_verifier.c
   - test/core/end2end/fake_resolver.c
-  - test/core/end2end/fixtures/http_proxy_fixture.c
+  - test/core/end2end/fixtures/http_proxy.c
   - test/core/end2end/fixtures/proxy.c
   - test/core/iomgr/endpoint_tests.c
   - test/core/util/debugger_macros.c
@@ -1214,20 +1214,6 @@ libs:
   - grpc++_codegen_base_src
   secure: false
   vs_project_guid: '{6EE56155-DF7C-4F6E-BFC4-F6F776BEB211}'
-- name: grpc_benchmark
-  build: test
-  language: c++
-  headers:
-  - test/cpp/microbenchmarks/fullstack_context_mutators.h
-  - test/cpp/microbenchmarks/fullstack_fixtures.h
-  - test/cpp/microbenchmarks/helpers.h
-  src:
-  - test/cpp/microbenchmarks/helpers.cc
-  deps:
-  - benchmark
-  - grpc++
-  - grpc_test_util
-  - grpc
 - name: grpc_cli_libs
   build: private
   language: c++
@@ -3045,7 +3031,6 @@ targets:
   src:
   - test/cpp/microbenchmarks/bm_call_create.cc
   deps:
-  - grpc_benchmark
   - benchmark
   - grpc++_test_util
   - grpc_test_util
@@ -3065,7 +3050,6 @@ targets:
   src:
   - test/cpp/microbenchmarks/bm_chttp2_hpack.cc
   deps:
-  - grpc_benchmark
   - benchmark
   - grpc++_test_util
   - grpc_test_util
@@ -3085,7 +3069,6 @@ targets:
   src:
   - test/cpp/microbenchmarks/bm_closure.cc
   deps:
-  - grpc_benchmark
   - benchmark
   - grpc++_test_util
   - grpc_test_util
@@ -3105,7 +3088,6 @@ targets:
   src:
   - test/cpp/microbenchmarks/bm_cq.cc
   deps:
-  - grpc_benchmark
   - benchmark
   - grpc++_test_util
   - grpc_test_util
@@ -3125,7 +3107,6 @@ targets:
   src:
   - test/cpp/microbenchmarks/bm_error.cc
   deps:
-  - grpc_benchmark
   - benchmark
   - grpc++_test_util
   - grpc_test_util
@@ -3139,85 +3120,12 @@ targets:
   - mac
   - linux
   - posix
-- name: bm_fullstack_streaming_ping_pong
+- name: bm_fullstack
   build: test
   language: c++
   src:
-  - test/cpp/microbenchmarks/bm_fullstack_streaming_ping_pong.cc
+  - test/cpp/microbenchmarks/bm_fullstack.cc
   deps:
-  - grpc_benchmark
-  - benchmark
-  - grpc++_test_util
-  - grpc_test_util
-  - grpc++
-  - grpc
-  - gpr_test_util
-  - gpr
-  args:
-  - --benchmark_min_time=0
-  excluded_poll_engines:
-  - poll
-  - poll-cv
-  platforms:
-  - mac
-  - linux
-  - posix
-  timeout_seconds: 1200
-- name: bm_fullstack_streaming_pump
-  build: test
-  language: c++
-  src:
-  - test/cpp/microbenchmarks/bm_fullstack_streaming_pump.cc
-  deps:
-  - grpc_benchmark
-  - benchmark
-  - grpc++_test_util
-  - grpc_test_util
-  - grpc++
-  - grpc
-  - gpr_test_util
-  - gpr
-  args:
-  - --benchmark_min_time=0
-  excluded_poll_engines:
-  - poll
-  - poll-cv
-  platforms:
-  - mac
-  - linux
-  - posix
-  timeout_seconds: 1200
-- name: bm_fullstack_trickle
-  build: test
-  language: c++
-  src:
-  - test/cpp/microbenchmarks/bm_fullstack_trickle.cc
-  deps:
-  - grpc_benchmark
-  - benchmark
-  - grpc++_test_util
-  - grpc_test_util
-  - grpc++
-  - grpc
-  - gpr_test_util
-  - gpr
-  args:
-  - --benchmark_min_time=0
-  excluded_poll_engines:
-  - poll
-  - poll-cv
-  platforms:
-  - mac
-  - linux
-  - posix
-  timeout_seconds: 1200
-- name: bm_fullstack_unary_ping_pong
-  build: test
-  language: c++
-  src:
-  - test/cpp/microbenchmarks/bm_fullstack_unary_ping_pong.cc
-  deps:
-  - grpc_benchmark
   - benchmark
   - grpc++_test_util
   - grpc_test_util
@@ -3241,7 +3149,6 @@ targets:
   src:
   - test/cpp/microbenchmarks/bm_metadata.cc
   deps:
-  - grpc_benchmark
   - benchmark
   - grpc_test_util
   - grpc
