+++ include/grpc/impl/codegen/sync.h
@@ -52,10 +52,6 @@
                                  provides no memory barriers.
  */
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 /* Platform-specific type declarations of gpr_mu and gpr_cv.   */
 #include <grpc/impl/codegen/port_platform.h>
 #include <grpc/impl/codegen/sync_generic.h>
@@ -66,10 +62,6 @@ extern "C" {
 #include <grpc/impl/codegen/sync_windows.h>
 #elif !defined(GPR_CUSTOM_SYNC)
 #error Unable to determine platform for sync
-#endif
-
-#ifdef __cplusplus
-}
 #endif
 
 #endif /* GRPC_IMPL_CODEGEN_SYNC_H */
